requirements.txt
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6-sip==13.8.0
SQLAlchemy==2.0.36
alembic==1.13.3
mutagen==1.47.0
python-dotenv==1.0.1
requests==2.32.3
